requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
anthropic==0.7.8
reportlab==4.0.7
pathlib2==2.3.7
gunicorn==21.2.0